move comp img to separate file
