+++ src/components/EmojiImg/emojiImg.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import emojiSrc from '../../lib/emojiSrc';
 import './style.css';
-import { useSetMissingEmoji } from '../../PickerContext';
+import Img from './img.jsx';
 
 const EmojiImg = ({ unified, shouldLoad = true, native = false, urlSrc }) => {
   return native ? (
@@ -18,37 +17,7 @@ const EmojiImg = ({ unified, shouldLoad = true, native = false, urlSrc }) => {
   );
 };
 
-function Img({ unified, shouldLoad, urlSrc }) {
-  const setMissingEmoji = useSetMissingEmoji();
-  const src = urlSrc ?? emojiSrc(unified);
-  //todo:erez should make more wise condition
-  if (urlSrc) {
-    return (
-      <img
-        className="emoji-img"
-        onError={() => setMissingEmoji(unified)}
-        src={urlSrc}
-      />
-    );
-  } else {
-    return (
-      <img
-        className="emoji-img"
-        onError={() => setMissingEmoji(unified)}
-        {...(shouldLoad && src)}
-      />
-    );
-  }
-}
-
 export default EmojiImg;
-
-Img.propTypes = {
-  unified: PropTypes.string,
-  shouldLoad: PropTypes.bool,
-  isCustom: PropTypes.bool,
-  urlSrc: PropTypes.string,
-};
 
 EmojiImg.propTypes = {
   unified: PropTypes.string,
